requirements/base.txt
# SemBench Framework - Shared Dependencies
# These are common deps needed by the SemBench framework itself
# (runners, evaluators, scenarios, plotting).

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.25.0
pyarrow>=14.0.0

# Scientific Computing & Evaluation
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
typing-extensions>=4.5.0

# Data Download & I/O
gdown>=4.7.0
requests>=2.31.0
aiohttp>=3.9.0
pypdf>=3.0.0
pdf2image>=1.16.0

# Database
duckdb

# Utilities
tqdm>=4.65.0
click>=8.1.0
natsort>=8.0.0
python-dateutil>=2.8.0
colorlog>=6.7.0
colorama>=0.4.6
tenacity>=8.2.0
dirtyjson>=1.0.8
tomli>=2.0.0
overrides>=7.4.0
jinja2>=3.1.0
openpyxl>=3.1.0

# ML / AI (shared across multiple systems)
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
accelerate>=0.20.0
safetensors>=0.3.0
tokenizers>=0.13.0
openai>=1.0.0
litellm>=1.0.0,<2.0.0
google-auth>=2.20.0
google-cloud-aiplatform>=1.38.0

# Evaluation
evaluate>=0.4.0
datasets>=2.14.0
cdlib>=0.4.0

# Testing
pytest>=7.4.0
